+++ src/alfasim_sdk/_internal/alfacase/case_to_alfacase.py
@@ -38,21 +38,7 @@ def format_list(values: List[Any], *, enable_flow_style: bool = False):
     For more details check:
     https://stackoverflow.com/questions/63364894/how-to-dump-only-lists-with-flow-style-with-pyyaml-or-ruamel-yaml
     """
-    import strictyaml
-
-    # The strictyaml version used by alfasim has ruamel vendored
-    # making it unable to work with the "non-vendored" ruamel version.
-    # The strictyaml version obtained with pip to run test on github
-    # uses a non vendored version of ruamel.
-    # By now we use by default here the vendored ruamel with a fallback
-    # to the original ruamel.
-    # When updating alfasim's strictyaml we probably should review this hack.
-    if hasattr(strictyaml, "ruamel"):
-        comments = strictyaml.ruamel.comments
-    else:
-        import ruamel
-
-        comments = ruamel.yaml.comments
+    from strictyaml.ruamel import comments
 
     retval = comments.CommentedSeq(values)
 
@@ -96,7 +82,7 @@ def _convert_value_to_valid_alfacase_format(
         )
 
     if isinstance(value, list) and all(
-        (isinstance(item, np.ndarray) and item.ndim == 1 for item in value)
+        isinstance(item, np.ndarray) and item.ndim == 1 for item in value
     ):
         return [
             format_list(
@@ -106,7 +92,7 @@ def _convert_value_to_valid_alfacase_format(
             for np_array in value
         ]
 
-    if isinstance(value, list) and all((isinstance(item, Array) for item in value)):
+    if isinstance(value, list) and all(isinstance(item, Array) for item in value):
         return [
             {"values": [str(i) for i in item.values], "unit": item.unit}
             for item in value